chore: Ignore the 'work_dir' directory in Sourcery configuration
The code changes include adding the 'work_dir/' directory to the list of paths ignored by Sourcery. This change ensures that the 'work_dir' directory is not processed by Sourcery.

Note: The commit message follows the conventional format of `<type>: <description>`. The type "chore" is used to indicate changes that are related to build processes, tools, or libraries.